Rename base hail table
