+++ v03_pipeline/lib/tasks/base/base_update_task.py
@@ -1,7 +1,7 @@
 import hail as hl
 
 from v03_pipeline.lib.misc.io import write
-from v03_pipeline.lib.tasks.base.base_hail_table_task import BaseHailTableTask
+from v03_pipeline.lib.tasks.base.base_hail_table import BaseHailTableTask
 
 
 class BaseUpdateTask(BaseHailTableTask):
